Make new and save/save_file files for layout package and re-order implementation to match interface
