+++ layout/layout.go
@@ -29,99 +29,112 @@ type Image struct {
 	refName    string // holds org.opencontainers.image.ref.name value
 }
 
-func NewImage(path string, ops ...ImageOption) (*Image, error) {
-	imageOpts := &options{}
-	for _, op := range ops {
-		if err := op(imageOpts); err != nil {
-			return nil, err
+// getters
+
+func (i *Image) Architecture() (string, error) {
+	cfg, err := i.Image.ConfigFile()
+	if err != nil {
+		return "", errors.Wrapf(err, "getting config file for image at path %q", i.path)
+	}
+	if cfg == nil {
+		return "", fmt.Errorf("missing config for image at path %q", i.path)
+	}
+	if cfg.Architecture == "" {
+		return "", fmt.Errorf("missing Architecture for image at path %q", i.path)
+	}
+	return cfg.Architecture, nil
+}
+
+func (i *Image) CreatedAt() (time.Time, error) {
+	configFile, err := i.Image.ConfigFile()
+	if err != nil {
+		return time.Time{}, errors.Wrapf(err, "getting createdAt time for image at path %q", i.path)
+	}
+	return configFile.Created.UTC(), nil
+}
+
+func (i *Image) Env(key string) (string, error) {
+	cfg, err := i.Image.ConfigFile()
+	if err != nil {
+		return "", errors.Wrapf(err, "getting config file for image at path %q", i.path)
+	}
+	if cfg == nil {
+		return "", fmt.Errorf("missing config for image at path %q", i.path)
+	}
+	for _, envVar := range cfg.Config.Env {
+		parts := strings.Split(envVar, "=")
+		if parts[0] == key {
+			return parts[1], nil
 		}
 	}
+	return "", nil
+}
 
-	platform := defaultPlatform()
-	if (imageOpts.platform != imgutil.Platform{}) {
-		platform = imageOpts.platform
+func (i *Image) Entrypoint() ([]string, error) {
+	cfg, err := i.Image.ConfigFile()
+	if err != nil {
+		return nil, errors.Wrapf(err, "getting config file for image at path %q", i.path)
 	}
+	if cfg == nil {
+		return nil, fmt.Errorf("missing config for image at path %q", i.path)
+	}
+	return cfg.Config.Entrypoint, nil
+}
 
-	image, err := emptyImage(platform)
+// Found tells whether the image exists in the repository by `Name()`.
+func (i *Image) Found() bool {
+	return ImageExists(i.path)
+}
+
+func ImageExists(path string) bool {
+	if !pathExists(path) {
+		return false
+	}
+	index := filepath.Join(path, "index.json")
+	if _, err := os.Stat(index); os.IsNotExist(err) {
+		return false
+	}
+	return true
+}
+
+func pathExists(path string) bool {
+	if path != "" {
+		if _, err := os.Stat(path); !os.IsNotExist(err) {
+			return true
+		}
+	}
+	return false
+}
+
+func (i *Image) GetAnnotateRefName() (string, error) {
+	return i.refName, nil
+}
+
+// GetLayer retrieves layer by diff id. Returns a reader of the uncompressed contents of the layer.
+// When the layers (notExistsLayer) came from a sparse image returns an empty reader
+func (i *Image) GetLayer(sha string) (io.ReadCloser, error) {
+	layers, err := i.Image.Layers()
 	if err != nil {
 		return nil, err
 	}
 
-	ri := &Image{
-		Image: image,
-		path:  path,
-	}
-
-	if imageOpts.prevImagePath != "" {
-		if err := processPreviousImageOption(ri, imageOpts.prevImagePath, platform); err != nil {
-			return nil, err
-		}
-	}
-
-	if imageOpts.baseImagePath != "" {
-		if err := processBaseImagePathOption(ri, imageOpts.baseImagePath, platform); err != nil {
-			return nil, err
-		}
-	} else if imageOpts.baseImage != nil {
-		if err := ri.mutateImage(imageOpts.baseImage); err != nil {
-			return nil, err
-		}
-	}
-
-	if imageOpts.createdAt.IsZero() {
-		ri.createdAt = imgutil.NormalizedDateTime
-	} else {
-		ri.createdAt = imageOpts.createdAt
-	}
-
-	return ri, nil
-}
-
-func processPreviousImageOption(ri *Image, prevImagePath string, platform imgutil.Platform) error {
-	prevImage, err := newV1Image(prevImagePath, platform)
+	layer, err := findLayerWithSha(layers, sha)
 	if err != nil {
-		return err
+		return nil, err
 	}
 
-	prevLayers, err := prevImage.Layers()
+	return layer.Uncompressed()
+}
+
+// Identifier
+// Each image's ID is given by the SHA256 hash of its configuration JSON. It is represented as a hexadecimal encoding of 256 bits,
+// e.g., sha256:a9561eb1b190625c9adb5a9513e72c4dedafc1cb2d4c5236c9a6957ec7dfd5a9.
+func (i *Image) Identifier() (imgutil.Identifier, error) {
+	hash, err := i.Image.Digest()
 	if err != nil {
-		return errors.Wrapf(err, "getting layers for previous image with path %q", prevImagePath)
+		return nil, errors.Wrapf(err, "getting identifier for image at path %q", i.path)
 	}
-
-	ri.prevLayers = prevLayers
-
-	return nil
-}
-
-func processBaseImagePathOption(ri *Image, baseImagePath string, platform imgutil.Platform) error {
-	baseImage, err := newV1Image(baseImagePath, platform)
-	if err != nil {
-		return err
-	}
-
-	return ri.mutateImage(baseImage)
-}
-
-func emptyImage(platform imgutil.Platform) (v1.Image, error) {
-	cfg := &v1.ConfigFile{
-		Architecture: platform.Architecture,
-		OS:           platform.OS,
-		OSVersion:    platform.OSVersion,
-		RootFS: v1.RootFS{
-			Type:    "layers",
-			DiffIDs: []v1.Hash{},
-		},
-	}
-	image := mutate.MediaType(empty.Image, types.OCIManifestSchema1)
-	image = mutate.ConfigMediaType(image, types.OCIConfigJSON)
-	return mutate.ConfigFile(image, cfg)
-}
-
-func defaultPlatform() imgutil.Platform {
-	return imgutil.Platform{
-		OS:           "linux",
-		Architecture: "amd64",
-	}
+	return newLayoutIdentifier(i.path, hash)
 }
 
 func (i *Image) Label(key string) (string, error) {
@@ -147,43 +160,43 @@ func (i *Image) Labels() (map[string]string, error) {
 	return cfg.Config.Labels, nil
 }
 
-func (i *Image) Env(key string) (string, error) {
-	cfg, err := i.Image.ConfigFile()
+// Layers overrides v1.Image Layers(), because we allow sparse image in OCI layout, sometimes some blobs
+// are missing. This method checks:
+// If there is data, return the layer
+// If there is no data, return a notExistsLayer
+func (i *Image) Layers() ([]v1.Layer, error) {
+	layers, err := i.Image.Layers()
 	if err != nil {
-		return "", errors.Wrapf(err, "getting config file for image at path %q", i.path)
+		return nil, err
 	}
-	if cfg == nil {
-		return "", fmt.Errorf("missing config for image at path %q", i.path)
-	}
-	for _, envVar := range cfg.Config.Env {
-		parts := strings.Split(envVar, "=")
-		if parts[0] == key {
-			return parts[1], nil
+
+	var retLayers []v1.Layer
+	for pos, layer := range layers {
+		if hasData(layer) {
+			retLayers = append(retLayers, layer)
+		} else {
+			cfg, err := i.Image.ConfigFile()
+			if err != nil {
+				return nil, err
+			}
+			diffID := cfg.RootFS.DiffIDs[pos]
+			retLayers = append(retLayers, &notExistsLayer{Layer: layer, diffID: diffID})
 		}
 	}
-	return "", nil
+	return retLayers, nil
 }
 
-func (i *Image) WorkingDir() (string, error) {
-	cfg, err := i.Image.ConfigFile()
-	if err != nil {
-		return "", errors.Wrapf(err, "getting config file for image at path %q", i.path)
-	}
-	if cfg == nil {
-		return "", fmt.Errorf("missing config for image at path %q", i.path)
-	}
-	return cfg.Config.WorkingDir, nil
+func hasData(layer v1.Layer) bool {
+	_, err := layer.Compressed()
+	return err == nil
 }
 
-func (i *Image) Entrypoint() ([]string, error) {
-	cfg, err := i.Image.ConfigFile()
-	if err != nil {
-		return nil, errors.Wrapf(err, "getting config file for image at path %q", i.path)
-	}
-	if cfg == nil {
-		return nil, fmt.Errorf("missing config for image at path %q", i.path)
-	}
-	return cfg.Config.Entrypoint, nil
+func (i *Image) ManifestSize() (int64, error) {
+	return i.Image.Size()
+}
+
+func (i *Image) Name() string {
+	return i.path
 }
 
 func (i *Image) OS() (string, error) {
@@ -211,18 +224,20 @@ func (i *Image) OSVersion() (string, error) {
 	return cfg.OSVersion, nil
 }
 
-func (i *Image) Architecture() (string, error) {
-	cfg, err := i.Image.ConfigFile()
+func (i *Image) TopLayer() (string, error) {
+	all, err := i.Image.Layers()
 	if err != nil {
-		return "", errors.Wrapf(err, "getting config file for image at path %q", i.path)
+		return "", err
 	}
-	if cfg == nil {
-		return "", fmt.Errorf("missing config for image at path %q", i.path)
+	if len(all) == 0 {
+		return "", fmt.Errorf("image at path %q has no layers", i.Name())
 	}
-	if cfg.Architecture == "" {
-		return "", fmt.Errorf("missing Architecture for image at path %q", i.path)
+	topLayer := all[len(all)-1]
+	hex, err := topLayer.DiffID()
+	if err != nil {
+		return "", err
 	}
-	return cfg.Architecture, nil
+	return hex.String(), nil
 }
 
 func (i *Image) Variant() (string, error) {
@@ -236,69 +251,45 @@ func (i *Image) Variant() (string, error) {
 	return cfg.Variant, nil
 }
 
-func (i *Image) Name() string {
-	return i.path
+func (i *Image) WorkingDir() (string, error) {
+	cfg, err := i.Image.ConfigFile()
+	if err != nil {
+		return "", errors.Wrapf(err, "getting config file for image at path %q", i.path)
+	}
+	if cfg == nil {
+		return "", fmt.Errorf("missing config for image at path %q", i.path)
+	}
+	return cfg.Config.WorkingDir, nil
+}
+
+// setters
+
+func (i *Image) AnnotateRefName(refName string) error {
+	i.refName = refName
+	return nil
 }
 
 func (i *Image) Rename(name string) {
 	i.path = name
 }
 
-// Found tells whether the image exists in the repository by `Name()`.
-func (i *Image) Found() bool {
-	return ImageExists(i.path)
-}
-
-// Identifier
-// Each image's ID is given by the SHA256 hash of its configuration JSON. It is represented as a hexadecimal encoding of 256 bits,
-// e.g., sha256:a9561eb1b190625c9adb5a9513e72c4dedafc1cb2d4c5236c9a6957ec7dfd5a9.
-func (i *Image) Identifier() (imgutil.Identifier, error) {
-	hash, err := i.Image.Digest()
-	if err != nil {
-		return nil, errors.Wrapf(err, "getting identifier for image at path %q", i.path)
-	}
-	return newLayoutIdentifier(i.path, hash)
-}
-
-func (i *Image) CreatedAt() (time.Time, error) {
+func (i *Image) SetArchitecture(architecture string) error {
 	configFile, err := i.Image.ConfigFile()
 	if err != nil {
-		return time.Time{}, errors.Wrapf(err, "getting createdAt time for image at path %q", i.path)
+		return err
 	}
-	return configFile.Created.UTC(), nil
+	configFile.Architecture = architecture
+	err = i.mutateConfigFile(i.Image, configFile)
+	return err
 }
 
-func (i *Image) Rebase(s string, image imgutil.Image) error {
-	return errors.New("not yet implemented")
-}
-
-func (i *Image) SetLabel(key string, val string) error {
+func (i *Image) SetCmd(cmd ...string) error {
 	configFile, err := i.Image.ConfigFile()
 	if err != nil {
 		return err
 	}
 	config := *configFile.Config.DeepCopy()
-	if config.Labels == nil {
-		config.Labels = map[string]string{}
-	}
-	config.Labels[key] = val
-	err = i.mutateConfig(i.Image, config)
-	if err != nil {
-		return errors.Wrapf(err, "set label key=%s value=%s", key, val)
-	}
-	return nil
-}
-
-func (i *Image) RemoveLabel(key string) error {
-	cfg, err := i.Image.ConfigFile()
-	if err != nil {
-		return errors.Wrapf(err, "getting config file for image at path %q", i.path)
-	}
-	if cfg == nil {
-		return fmt.Errorf("missing config for image at path %q", i.path)
-	}
-	config := *cfg.Config.DeepCopy()
-	delete(config.Labels, key)
+	config.Cmd = cmd
 	err = i.mutateConfig(i.Image, config)
 	return err
 }
@@ -329,17 +320,6 @@ func (i *Image) SetEnv(key string, val string) error {
 	return err
 }
 
-func (i *Image) SetWorkingDir(dir string) error {
-	configFile, err := i.Image.ConfigFile()
-	if err != nil {
-		return err
-	}
-	config := *configFile.Config.DeepCopy()
-	config.WorkingDir = dir
-	err = i.mutateConfig(i.Image, config)
-	return err
-}
-
 func (i *Image) SetEntrypoint(ep ...string) error {
 	configFile, err := i.Image.ConfigFile()
 	if err != nil {
@@ -351,15 +331,21 @@ func (i *Image) SetEntrypoint(ep ...string) error {
 	return err
 }
 
-func (i *Image) SetCmd(cmd ...string) error {
+func (i *Image) SetLabel(key string, val string) error {
 	configFile, err := i.Image.ConfigFile()
 	if err != nil {
 		return err
 	}
 	config := *configFile.Config.DeepCopy()
-	config.Cmd = cmd
+	if config.Labels == nil {
+		config.Labels = map[string]string{}
+	}
+	config.Labels[key] = val
 	err = i.mutateConfig(i.Image, config)
-	return err
+	if err != nil {
+		return errors.Wrapf(err, "set label key=%s value=%s", key, val)
+	}
+	return nil
 }
 
 func (i *Image) SetOS(osVal string) error {
@@ -382,16 +368,6 @@ func (i *Image) SetOSVersion(osVersion string) error {
 	return err
 }
 
-func (i *Image) SetArchitecture(architecture string) error {
-	configFile, err := i.Image.ConfigFile()
-	if err != nil {
-		return err
-	}
-	configFile.Architecture = architecture
-	err = i.mutateConfigFile(i.Image, configFile)
-	return err
-}
-
 func (i *Image) SetVariant(variant string) error {
 	configFile, err := i.Image.ConfigFile()
 	if err != nil {
@@ -402,37 +378,18 @@ func (i *Image) SetVariant(variant string) error {
 	return err
 }
 
-func (i *Image) TopLayer() (string, error) {
-	all, err := i.Image.Layers()
+func (i *Image) SetWorkingDir(dir string) error {
+	configFile, err := i.Image.ConfigFile()
 	if err != nil {
-		return "", err
+		return err
 	}
-	if len(all) == 0 {
-		return "", fmt.Errorf("image at path %q has no layers", i.Name())
-	}
-	topLayer := all[len(all)-1]
-	hex, err := topLayer.DiffID()
-	if err != nil {
-		return "", err
-	}
-	return hex.String(), nil
+	config := *configFile.Config.DeepCopy()
+	config.WorkingDir = dir
+	err = i.mutateConfig(i.Image, config)
+	return err
 }
 
-// GetLayer retrieves layer by diff id. Returns a reader of the uncompressed contents of the layer.
-// When the layers (notExistsLayer) came from a sparse image returns an empty reader
-func (i *Image) GetLayer(sha string) (io.ReadCloser, error) {
-	layers, err := i.Image.Layers()
-	if err != nil {
-		return nil, err
-	}
-
-	layer, err := findLayerWithSha(layers, sha)
-	if err != nil {
-		return nil, err
-	}
-
-	return layer.Uncompressed()
-}
+// modifiers
 
 // AddLayer adds an uncompressed tarred layer to the image
 func (i *Image) AddLayer(path string) error {
@@ -443,10 +400,55 @@ func (i *Image) AddLayer(path string) error {
 	return i.addOCILayer(layer)
 }
 
+// addOCILayer appends the provided layer with media type application/vnd.oci.image.layer.v1.tar+gzip
+func (i *Image) addOCILayer(layer v1.Layer) error {
+	additions := layersAddendum([]v1.Layer{layer})
+	image, err := mutate.Append(i.Image, additions...)
+	if err != nil {
+		return errors.Wrap(err, "add layer")
+	}
+	return i.mutateImage(image)
+}
+
+// layersAddendum creates an Addendum array with the given layers
+// and 'application/vnd.oci.image.layer.v1.tar+gzip' media type
+func layersAddendum(layers []v1.Layer) []mutate.Addendum {
+	additions := make([]mutate.Addendum, 0)
+	for _, layer := range layers {
+		additions = append(additions, mutate.Addendum{
+			MediaType: types.OCILayer,
+			Layer:     layer,
+		})
+	}
+	return additions
+}
+
 func (i *Image) AddLayerWithDiffID(path, diffID string) error {
 	// this is equivalent to AddLayer in the layout case
 	// it exists to provide optimize performance for local images
 	return i.AddLayer(path)
+}
+
+func (i *Image) Delete() error {
+	return os.RemoveAll(i.path)
+}
+
+func (i *Image) Rebase(s string, image imgutil.Image) error {
+	return errors.New("not yet implemented")
+}
+
+func (i *Image) RemoveLabel(key string) error {
+	cfg, err := i.Image.ConfigFile()
+	if err != nil {
+		return errors.Wrapf(err, "getting config file for image at path %q", i.path)
+	}
+	if cfg == nil {
+		return fmt.Errorf("missing config for image at path %q", i.path)
+	}
+	config := *cfg.Config.DeepCopy()
+	delete(config.Labels, key)
+	err = i.mutateConfig(i.Image, config)
+	return err
 }
 
 func (i *Image) ReuseLayer(sha string) error {
@@ -456,6 +458,8 @@ func (i *Image) ReuseLayer(sha string) error {
 	}
 	return i.addOCILayer(layer)
 }
+
+// helpers
 
 func findLayerWithSha(layers []v1.Layer, diffID string) (v1.Layer, error) {
 	for _, layer := range layers {
@@ -468,207 +472,6 @@ func findLayerWithSha(layers []v1.Layer, diffID string) (v1.Layer, error) {
 		}
 	}
 	return nil, fmt.Errorf("previous image did not have layer with diff id %q", diffID)
-}
-
-func (i *Image) Save(additionalNames ...string) error {
-	return i.SaveAs(i.Name(), additionalNames...)
-}
-
-// SaveAs ignores the image `Name()` method and saves the image according to name & additional names provided to this method
-func (i *Image) SaveAs(name string, additionalNames ...string) error {
-	err := i.mutateCreatedAt(i.Image, v1.Time{Time: i.createdAt})
-	if err != nil {
-		return errors.Wrap(err, "set creation time")
-	}
-
-	cfg, err := i.Image.ConfigFile()
-	if err != nil {
-		return errors.Wrap(err, "get image config")
-	}
-	cfg = cfg.DeepCopy()
-
-	layers, err := i.Image.Layers()
-	if err != nil {
-		return errors.Wrap(err, "get image layers")
-	}
-	cfg.History = make([]v1.History, len(layers))
-	for j := range cfg.History {
-		cfg.History[j] = v1.History{
-			Created: v1.Time{Time: i.createdAt},
-		}
-	}
-
-	cfg.DockerVersion = ""
-	cfg.Container = ""
-	err = i.mutateConfigFile(i.Image, cfg)
-	if err != nil {
-		return errors.Wrap(err, "zeroing history")
-	}
-
-	var diagnostics []imgutil.SaveDiagnostic
-	annotations := ImageRefAnnotation(i.refName)
-	pathsToSave := append([]string{name}, additionalNames...)
-	for _, path := range pathsToSave {
-		// initialize image path
-		path, err := Write(path, empty.Index)
-		if err != nil {
-			return err
-		}
-
-		err = path.AppendImage(i.Image, WithAnnotations(annotations))
-		if err != nil {
-			diagnostics = append(diagnostics, imgutil.SaveDiagnostic{ImageName: i.Name(), Cause: err})
-		}
-	}
-
-	if len(diagnostics) > 0 {
-		return imgutil.SaveError{Errors: diagnostics}
-	}
-
-	return nil
-}
-
-func (i *Image) SaveFile() (string, error) {
-	// TODO issue https://github.com/buildpacks/imgutil/issues/170
-	return "", errors.New("not yet implemented")
-}
-
-func (i *Image) Delete() error {
-	return os.RemoveAll(i.path)
-}
-
-func (i *Image) ManifestSize() (int64, error) {
-	return i.Image.Size()
-}
-
-// Layers overrides v1.Image Layers(), because we allow sparse image in OCI layout, sometimes some blobs
-// are missing. This method checks:
-// If there is data, return the layer
-// If there is no data, return a notExistsLayer
-func (i *Image) Layers() ([]v1.Layer, error) {
-	layers, err := i.Image.Layers()
-	if err != nil {
-		return nil, err
-	}
-
-	var retLayers []v1.Layer
-	for pos, layer := range layers {
-		if hasData(layer) {
-			retLayers = append(retLayers, layer)
-		} else {
-			cfg, err := i.Image.ConfigFile()
-			if err != nil {
-				return nil, err
-			}
-			diffID := cfg.RootFS.DiffIDs[pos]
-			retLayers = append(retLayers, &notExistsLayer{Layer: layer, diffID: diffID})
-		}
-	}
-	return retLayers, nil
-}
-
-func (i *Image) AnnotateRefName(refName string) error {
-	i.refName = refName
-	return nil
-}
-
-func (i *Image) GetAnnotateRefName() (string, error) {
-	return i.refName, nil
-}
-
-func ImageExists(path string) bool {
-	if !pathExists(path) {
-		return false
-	}
-	index := filepath.Join(path, "index.json")
-	if _, err := os.Stat(index); os.IsNotExist(err) {
-		return false
-	}
-	return true
-}
-
-func hasData(layer v1.Layer) bool {
-	_, err := layer.Compressed()
-	return err == nil
-}
-
-func pathExists(path string) bool {
-	if path != "" {
-		if _, err := os.Stat(path); !os.IsNotExist(err) {
-			return true
-		}
-	}
-	return false
-}
-
-// newV1Image creates a layout image from the given path.
-//   - If a ImageIndex for multiples platforms exists, then it will try to select the image
-//     according to the platform provided
-//   - If the image does not exist, then an empty image is returned
-func newV1Image(path string, platform imgutil.Platform) (v1.Image, error) {
-	var (
-		image  v1.Image
-		layout Path
-		err    error
-	)
-
-	if ImageExists(path) {
-		layout, err = FromPath(path)
-		if err != nil {
-			return nil, errors.Wrap(err, "loading layout from path new")
-		}
-
-		index, err := layout.ImageIndex()
-		if err != nil {
-			return nil, errors.Wrap(err, "reading index")
-		}
-
-		image, err = imageFromIndex(index, platform)
-		if err != nil {
-			return nil, errors.Wrap(err, "getting image from index")
-		}
-	} else {
-		image, err = emptyImage(platform)
-		if err != nil {
-			return nil, errors.Wrap(err, "initializing empty image")
-		}
-	}
-	return &Image{
-		Image: image,
-		path:  path,
-	}, nil
-}
-
-// imageFromIndex creates a v1.Image from the given Image Index, selecting the image manifest
-// that matches the given OS and architecture.
-func imageFromIndex(index v1.ImageIndex, platform imgutil.Platform) (v1.Image, error) {
-	indexManifest, err := index.IndexManifest()
-	if err != nil {
-		return nil, err
-	}
-
-	if len(indexManifest.Manifests) == 0 {
-		return nil, errors.New("no underlyingImage indexManifest found")
-	}
-
-	manifest := indexManifest.Manifests[0]
-	if len(indexManifest.Manifests) > 1 {
-		// Find based on platform (os/arch)
-		for _, m := range indexManifest.Manifests {
-			if m.Platform.OS == platform.OS && m.Platform.Architecture == platform.OS {
-				manifest = m
-				break
-			}
-		}
-		return nil, fmt.Errorf("manifest matching platform %v not found", platform)
-	}
-
-	image, err := index.Image(manifest.Digest)
-	if err != nil {
-		return nil, err
-	}
-
-	return image, nil
 }
 
 // mutateConfig mutates the provided v1.Image to have the provided v1.Config and wraps the result
@@ -685,16 +488,6 @@ func (i *Image) mutateConfig(base v1.Image, config v1.Config) error {
 // into a layout.Image (requires for override methods like Layers()
 func (i *Image) mutateConfigFile(base v1.Image, configFile *v1.ConfigFile) error {
 	image, err := mutate.ConfigFile(base, configFile)
-	if err != nil {
-		return err
-	}
-	return i.mutateImage(image)
-}
-
-// mutateCreatedAt mutates the provided v1.Image to have the provided v1.Time and wraps the result
-// into a layout.Image (requires for override methods like Layers()
-func (i *Image) mutateCreatedAt(base v1.Image, created v1.Time) error {
-	image, err := mutate.CreatedAt(i.Image, v1.Time{Time: i.createdAt})
 	if err != nil {
 		return err
 	}
@@ -722,16 +515,6 @@ func (i *Image) mutateImage(base v1.Image) error {
 		}
 	}
 	return nil
-}
-
-// addOCILayer appends the provided layer with media type application/vnd.oci.image.layer.v1.tar+gzip
-func (i *Image) addOCILayer(layer v1.Layer) error {
-	additions := layersAddendum([]v1.Layer{layer})
-	image, err := mutate.Append(i.Image, additions...)
-	if err != nil {
-		return errors.Wrap(err, "add layer")
-	}
-	return i.mutateImage(image)
 }
 
 // validMediaTypes returns true if media types present in the manifest are the ones defined by the OCI spec
@@ -769,17 +552,4 @@ func overrideMediaTypes(base v1.Image) (v1.Image, error) {
 	}
 
 	return image, nil
-}
-
-// layersAddendum creates an Addendum array with the given layers
-// and 'application/vnd.oci.image.layer.v1.tar+gzip' media type
-func layersAddendum(layers []v1.Layer) []mutate.Addendum {
-	additions := make([]mutate.Addendum, 0)
-	for _, layer := range layers {
-		additions = append(additions, mutate.Addendum{
-			MediaType: types.OCILayer,
-			Layer:     layer,
-		})
-	}
-	return additions
 }
